Imported another printf library
Don't feel like spending a bunch of time fiddling with it in a kernel
environment, maybe in another repo in the future.
Moved all printfs to new functions. Disabled floating point support
since we haven't enabled it yet in the kernel. Updated some formatting
and some warnings to try to cut down on extrenious warnings.

+++ HeliOS/kernel/screen.c
@@ -8,31 +8,33 @@ extern char _binary_fonts_font_psf_end;
 uint16_t* unicode = NULL;
 
 static struct screen_info sc = {
-    .cx = 0,
-    .cy = 0,
-    .fgc = 0xFFFFFF,
-    .bgc = 0x000000,
+	.cx = 0,
+	.cy = 0,
+	.fgc = 0xFFFFFF,
+	.bgc = 0x000000,
 };
 
 /* import our font that's in the object file we've created above */
 extern char _binary_font_start[];
 
-static void screen_putchar_at(unsigned short int c, int cx, int cy, uint32_t fg, uint32_t bg);
+static void screen_putchar_at(unsigned short int c, int cx, int cy, uint32_t fg,
+			      uint32_t bg);
 static void scroll();
 
 // TODO: pass through framebuffer and such
-void screen_init(struct limine_framebuffer* fb, uint32_t fg_color, uint32_t bg_color)
+void screen_init(struct limine_framebuffer* fb, uint32_t fg_color,
+		 uint32_t bg_color)
 {
-    // TODO: properly init psf, though the one im using currently isnt unicode
+	// TODO: properly init psf, though the one im using currently isnt unicode
 
-    sc.cx = 0;
-    sc.cy = 0;
-    sc.fgc = fg_color;
-    sc.bgc = bg_color;
-    sc.fb = fb;
-    sc.fb_buffer = (char*)fb->address;
-    sc.scanline = fb->pitch;
-    sc.font = (PSF_font*)&_binary_fonts_font_psf_start;
+	sc.cx = 0;
+	sc.cy = 0;
+	sc.fgc = fg_color;
+	sc.bgc = bg_color;
+	sc.fb = fb;
+	sc.fb_buffer = (char*)fb->address;
+	sc.scanline = fb->pitch;
+	sc.font = (PSF_font*)&_binary_fonts_font_psf_start;
 }
 
 /**
@@ -42,8 +44,8 @@ void screen_init(struct limine_framebuffer* fb, uint32_t fg_color, uint32_t bg_c
  */
 void set_color(uint32_t fg, uint32_t bg)
 {
-    sc.fgc = fg;
-    sc.bgc = bg;
+	sc.fgc = fg;
+	sc.bgc = bg;
 }
 
 /**
@@ -52,10 +54,10 @@ void set_color(uint32_t fg, uint32_t bg)
  */
 void screen_putstring(const char* s)
 {
-    while (*s) {
-        screen_putchar(*s);
-        s++;
-    }
+	while (*s) {
+		screen_putchar(*s);
+		s++;
+	}
 }
 
 /**
@@ -72,25 +74,25 @@ void screen_putstring(const char* s)
  */
 void screen_putchar(char c)
 {
-    switch (c) {
-    case '\n':
-        sc.cy++;
-        sc.cx = 0;
-        break;
-    case '\b':
-        if (sc.cx == 0) break;
-        screen_putchar_at(' ', --sc.cx, sc.cy, sc.fgc, sc.bgc);
-    default:
-        screen_putchar_at(c, sc.cx++, sc.cy, sc.fgc, sc.bgc);
-    }
-    if (sc.cx >= sc.scanline) {
-        sc.cx = 0;
-        sc.cy++;
-    }
-    if (sc.cy >= sc.fb->height / sc.font->height) {
-        scroll();
-        sc.cy = sc.cy - 2;
-    }
+	switch (c) {
+	case '\n':
+		sc.cy++;
+		sc.cx = 0;
+		break;
+	case '\b':
+		if (sc.cx == 0) break;
+		screen_putchar_at(' ', --sc.cx, sc.cy, sc.fgc, sc.bgc);
+	default:
+		screen_putchar_at(c, sc.cx++, sc.cy, sc.fgc, sc.bgc);
+	}
+	if (sc.cx >= (size_t)sc.scanline) {
+		sc.cx = 0;
+		sc.cy++;
+	}
+	if (sc.cy >= sc.fb->height / sc.font->height) {
+		scroll();
+		sc.cy = sc.cy - 2;
+	}
 }
 
 /**
@@ -106,21 +108,22 @@ void screen_putchar(char c)
  */
 static void scroll()
 {
-    uint32_t s_width = sc.fb->width;
-    uint32_t char_height = sc.font->height;
+	uint32_t s_width = sc.fb->width;
+	uint32_t char_height = sc.font->height;
 
-    uint32_t scroll_height = (sc.fb->height / sc.font->height) - 1;
-    uint64_t copy_size = char_height * s_width * (sc.fb->bpp / 8);
+	uint32_t scroll_height = (sc.fb->height / sc.font->height) - 1;
+	uint64_t copy_size = char_height * s_width * (sc.fb->bpp / 8);
 
-    void* addr = sc.fb->address;
+	void* addr = sc.fb->address;
 
-    // Move everything up
-    memcpy(addr, (void*)((uint64_t)addr + copy_size), copy_size * scroll_height);
+	// Move everything up
+	memcpy(addr, (void*)((uint64_t)addr + copy_size),
+	       copy_size * scroll_height);
 
-    // Clear the last row
-    uint64_t offset = copy_size * scroll_height;
-    void* last_row = (void*)((uint64_t)addr + offset);
-    memset(last_row, sc.bgc, copy_size);
+	// Clear the last row
+	uint64_t offset = copy_size * scroll_height;
+	void* last_row = (void*)((uint64_t)addr + offset);
+	memset(last_row, sc.bgc, copy_size);
 }
 
 /**
@@ -136,38 +139,43 @@ static void scroll()
  * @param fg The foreground color (e.g., 0xFFFFFF for white).
  * @param bg The background color (e.g., 0x000000 for black).
  */
-static void screen_putchar_at(unsigned short int c, int cx, int cy, uint32_t fg, uint32_t bg)
+static void screen_putchar_at(unsigned short int c, int cx, int cy, uint32_t fg,
+			      uint32_t bg)
 {
-    /* cast the address to PSF header struct */
-    PSF_font* font = (PSF_font*)&_binary_fonts_font_psf_start;
-    /* we need to know how many bytes encode one row */
-    int bytesperline = (font->width + 7) / 8;
-    /* unicode translation */
-    if (unicode != NULL) {
-        c = unicode[c];
-    }
-    /* get the glyph for the character. If there's no
+	/* cast the address to PSF header struct */
+	PSF_font* font = (PSF_font*)&_binary_fonts_font_psf_start;
+	/* we need to know how many bytes encode one row */
+	int bytesperline = (font->width + 7) / 8;
+	/* unicode translation */
+	if (unicode != NULL) {
+		c = unicode[c];
+	}
+	/* get the glyph for the character. If there's no
        glyph for a given character, we'll display the first glyph. */
-    unsigned char* glyph = (unsigned char*)&_binary_fonts_font_psf_start + font->headersize
-        + (c > 0 && c < font->numglyph ? c : 0) * font->bytesperglyph;
-    /* calculate the upper left corner on screen where we want to display.
+	unsigned char* glyph =
+		(unsigned char*)&_binary_fonts_font_psf_start +
+		font->headersize +
+		(c > 0 && c < font->numglyph ? c : 0) * font->bytesperglyph;
+	/* calculate the upper left corner on screen where we want to display.
        we only do this once, and adjust the offset later. This is faster. */
-    int offs = (cy * font->height * sc.scanline) + (cx * (font->width + 1) * sizeof(PIXEL));
-    /* finally display pixels according to the bitmap */
-    int x, y, line, mask;
-    for (y = 0; y < font->height; y++) {
-        /* save the starting position of the line */
-        line = offs;
-        mask = 1 << (font->width - 1);
-        /* display a row */
-        for (x = 0; x < font->width; x++) {
-            *((PIXEL*)(sc.fb_buffer + line)) = *((unsigned int*)glyph) & mask ? fg : bg;
-            /* adjust to the next pixel */
-            mask >>= 1;
-            line += sizeof(PIXEL);
-        }
-        /* adjust to the next line */
-        glyph += bytesperline;
-        offs += sc.scanline;
-    }
+	int offs = (cy * font->height * sc.scanline) +
+		   (cx * (font->width + 1) * sizeof(PIXEL));
+	/* finally display pixels according to the bitmap */
+	int x, y, line, mask;
+	for (y = 0; (uint32_t)y < font->height; y++) {
+		/* save the starting position of the line */
+		line = offs;
+		mask = 1 << (font->width - 1);
+		/* display a row */
+		for (x = 0; (uint32_t)x < font->width; x++) {
+			*((PIXEL*)(sc.fb_buffer + line)) =
+				*((unsigned int*)glyph) & mask ? fg : bg;
+			/* adjust to the next pixel */
+			mask >>= 1;
+			line += sizeof(PIXEL);
+		}
+		/* adjust to the next line */
+		glyph += bytesperline;
+		offs += sc.scanline;
+	}
 }
